pyproject.toml
[tool.poetry]
name = "spotify-playlist-importer"
version = "0.1.0"
description = ""
authors = ["ricklon <rianders@docs.rutgers.edu>"]
readme = "README.md"
packages = [{include = "spotify_playlist_importer"}]

[tool.poetry.dependencies]
python = "^3.9"
gradio = "^3.21.0"
spotipy = "^2.22.1"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
